refactor: drop is_valid_url + SourceNotFound deprecation shims
pipenv's stable API is the CLI, not the Python module surface
(T_C.3 §9 sign-off). A DeprecationWarning shim for an internal-only
re-export protects nothing — it just delays the cleanup. Delete both
shims outright, update the two real internal callers, and remove
the shim-only unit test.

This resolves T_A.4 (previously "held for release"); the release-gate
rationale is moot when there is no stable Python API to gate against.
Folds in the parallel cleanup of the SourceNotFound re-export that
T_D.2 left on Project for the same purported deprecation window.

- pipenv/utils/fileutils.py: delete is_valid_url shim + the
  ``_is_valid_url`` alias import; intra-module caller already used
  the canonical name.
- pipenv/project.py: drop SourceNotFound from the sources import.
- pipenv/utils/indexes.py: import SourceNotFound directly from
  pipenv.utils.sources (the only non-test internal caller).
- pipenv/utils/sources.py: trim the docstring reference to the
  removed re-export.
- tests/unit/test_utils.py: delete the shim-only deprecation test.
- tests/unit/test_sources.py: import SourceNotFound from
  pipenv.utils.sources.

650 unit tests pass.
